[include & src] Everything related to the contructors need to be defined in the h file...

+++ src/main.cpp
@@ -2,6 +2,6 @@
 
 
 int main(int argc, char** argv){
-    nn<f64> nn;
+    nn<f32> nn;
     return 0;
 }
